+++ spec/slack-sup/events/member_left_channel_spec.rb
@@ -9,9 +9,8 @@ describe 'events/member_left_channel' do
     {
       type: 'member_left_channel',
       team: team.team_id,
-      user: team.bot_user_id,
       channel: 'channel_id',
-      inviter: 'iviter_id'
+      channel_type: 'C'
     }
   end
 
@@ -19,17 +18,5 @@ describe 'events/member_left_channel' do
     post '/api/slack/event', event_envelope
     expect(last_response.status).to eq 201
     expect(JSON.parse(last_response.body)).to eq('ok' => true)
-  end
-
-  context 'with a channel' do
-    let!(:channel) { Fabricate(:channel, team:, channel_id: event[:channel]) }
-
-    it 'removes bot from channel' do
-      expect(channel.enabled).to be true
-      post '/api/slack/event', event_envelope
-      expect(last_response.status).to eq 201
-      expect(JSON.parse(last_response.body)).to eq('ok' => true)
-      expect(channel.reload.enabled).to be false
-    end
   end
 end