+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### Changelog
 
+* 2025/09/01: Fix: removing bot from channel doesn't disable S'Up - [@dblock](https://github.com/dblock).
 * 2025/05/04: Visibly handle vacationing users - [@dblock](https://github.com/dblock).
 * 2025/02/15: Added `data [n|all]` and only export the last round by default - [@dblock](https://github.com/dblock).
 * 2025/02/04: Added `admins` that shows team and channel admins - [@dblock](https://github.com/dblock).